eth_tasks - debug print taken out
init.c - line sof code previously use for debug and commented afterwords
taken out, as part of a code cleanup

iop.xml - The MAC adress of the iop_example had to be changed. A bug was
found on the board that does not allow the user to set the board MAC
address with the first two bytes different from zero.