Fix for retl bug. Possible stack pointer fix for thread context switch, not fully tested. Cache fixes.
RETL instruction needs two delay slots when returning from trap, so an
extra instruction isn't performed. This can be seen when when the RETT
instruction is followed by RETL instruction. The instruction on the RETT
delay slot is skipped and the next intruction performed on the delay
slot.
Stack pointer on thread context switch was being set by RTEMS to the
bottom of an interrupt stack frame, so that when the aprtition context
switch occured momory area relevant to the application waws overwritten.
This issues is being further investigated as this might not be a
definite fix.

refs #112319