Corrects the type definitions on air_arch.h
The air_uptr_t is now a pointer to an unsigned int. Due to the change, the ADDR_ALIGN and air_uptr_t use in the core was updated to work in ARM and SPARC definitions.

Also removes the SPARC function declaration in air_arch.h

refs #134456

Changes to be committed:
	modified:   air/libs/libtest/include/libtest.h
	modified:   air/pmk/arch/arm/include/air_arch.h
	modified:   air/pmk/core/include/segregation.h
	modified:   air/pmk/core/loader.c
	modified:   air/pmk/core/segregation.c

[OC] 2018/09/05