LIBIOP: UART driver
It is now possible to receive and transmit a message through UART.
The message received is broken into strings of seemingly random length.
TODO: This length should be the DataBytes defined by the user on the iop.xml file.
TODO: After a few cycles the program stops running (Data Abort)

refs #194753