LIBIOP: UART driver
The time_to_live field is obtained from the iop.xml file.
If the txFIFO is full before all of the data is transmitted, the part of the message that wasn't sent is preppended to the sendqueue and the program stays inside the loop until the sendqueue is empty.

TODO: When the FIFO is full, the program should break the writer task loop so the IOP can run the other tasks. With the current solution for the time_to_live field (in which it refers to the number of periods, not a time interval), the data in the sendqueue might be lost in the pre-dispatcher. This should be fixed before a 'break' is added in the writer task when the FIFO is full.

refs #194753