Added the first GR740 test
The first GR740 test (hello_world) to be run under laysim-gr740 was
added. Test runner script was also added (do_gr740_laysim.bash).
The test is currently running under Dynamic Binary Translator (DBT)
mode, in order to be faster. This could be changed to a more accurate
mode if errors are encountered.