Added Ethernet test.
Added a first version of the Ethernet CI/CD test. A couple of CI/CD
scripts had to be remade (new versions were added), and a host script
was added to accommodate having to run different parts of the test on
the board and host PC.

As of now, the only way the test fails is if the host receives no
messages from the board. It's assumed that if board-to-host communication
is working, host-to-board communication is also functional. Test logic
to be altered if this turns out not to be the case.