GPIO Driver: Configurator update.
Updated the relevant configurator files to support a GPIO IOP driver.
Adding a GPIO Port to an application requires two fields: Pin#, and
whether the pin is writable or not. My reasoning with this is so that
several partitions can access the same pin to read it. However, only those
configured to write can do it.