Changed Read/Write attributes to device header instead of route header.
Instead of declaring if a GPIO pin is to be read or written in the route
header, the device header in the configuration should have a write mask
(32 bits, right to left, 0 if read-only, 1 otherwise). This will be used
in the initialization routine of the GPIO bank. Also limited the number of
GPIO pins to be used to 32.