Virtual exception handling corrections
The goal is to let RTEMS do the context restoring for us so that we only need one context saved at a time per partition (and avoid complicated operations in exception.S)
Added auxiliary resgisters for mode virtualization in core context.
Adapted exception.S and syscalls to use these registers.